CrateDB: Add support for SQLRecordManager
It does not work, because this subsystem uses composite unique keys in
combination with an `ON CONFLICT DO UPDATE` operation, on behalf of the
model entity definition `UpsertionRecord`.


Because the composite uniqueness constraint is currently being emulated
already, it can't also emulate ON CONFLICT behaviour on top easily.

__table_args__ = (
    UniqueConstraint("key", "namespace", name="uix_key_namespace"),
    Index("ix_key_namespace", "key", "namespace"),
)

stmt = insert_stmt.on_conflict_do_update(
    [UpsertionRecord.key, UpsertionRecord.namespace],
    ...
)